requirements.txt
pandas
pyserial
PyQt5
numpy
telemetrix
nmea-parser
folium
pynmeagps